Drive the remodeled advisory-cluster callees toward ground truth (five more 100% matches)
Follow-up to f67c1e06, shaping the callees surfaced by the advisory
scoring sweep:

- TOcean::FindMapActionContextContainingNodeByIndex (0x564570,
  48% -> 100%): the containment scan is a pair of inlines on
  TZoneSecondaryNeighborStretch — T* FindEntry(T) (unsigned indexed
  for-loop returning &Data()[index]) and bool ContainsEntry (SETNE
  materialization). Declared on the concrete stretch subclass because
  MSVC500 eagerly instantiates template members (stretch<Seapoint> has
  no operator==).
- TArmyMgr::ComputeWeightedNeighborLinkScoreForNodeIndex (0x4a5aa0,
  67% -> 100%): int param truncated to a short local, chain=0
  fallthrough instead of an early return, and the 0x6955f0 weight table
  promoted to a real global
  (g_anWeightedNeighborUnitScoreByType_006955F0[32]) instead of a
  reinterpret_cast'd raw address.
- TGreatPower::SumNavyOrderPriorityForNation[AndNodeType] (0x4e0460 /
  0x4e04b0, both -> 100%): ComputeOrderNodeCompositeEconomicScore is a
  real __thiscall TShip method (callsites load ecx = node, no pushes);
  its body's SignedMod100 bit-twiddle was invented — the original is a
  plain short/100 and (x+w*10+5)/10 division chain (0x550b60
  20% -> 68%, residual FP-free scheduling wobble).
- TZone::ComputeMapActionContextNodeValueAverage (0x55f140,
  61% -> 95%): TPortZone::field48 retyped to its real short (every
  original reader is movsx word; casts dropped at all 7 sites),
  ownerTag as int local, single merged return-0 tail.
- TDiplomacyMgr::BuildRelationshipListSlot88 (0x4f1f70): materialized
  bool for the encodedNationSlot == -1 test; residual is register
  naming only.

Verified: just build green, just gates green, just test 109/109,
stats +0.02pp (12 small drops are register wobble in untouched
functions, baseline updated). Heuristics notes 116-117 added.

+++ include/game/TArmyMgr.h
@@ -126,7 +126,7 @@ public:
   // Real __thiscall on the TArmyMgr singleton (ret 4; both original callsites,
   // 0x004d8390 and 0x004d83c0, load g_pMapContextActionManager into ecx); `this`
   // is unused by the body.
-  int ComputeWeightedNeighborLinkScoreForNodeIndex(short nodeIndex);
+  int ComputeWeightedNeighborLinkScoreForNodeIndex(int nodeIndex);
 
   // Object size 0x3a8 confirmed by RTTI. Only +0x31c is confirmed so far (read from three
   // independent call sites: TArmyPlacard::HandleEvent, TArmyToolbar's equivalent, and
